+++ source/sounds.hpp
@@ -1,10 +1,17 @@
-// This work is licensed under the terms of the MIT license.
-// For a copy, see <https://opensource.org/licenses/MIT>.
+///
+/// This file is part of the SoundBox Project.
+/// Author: Graham Smith
+///
+/// This work is licensed under the terms of the MIT license.
+/// For a copy, see <https://opensource.org/licenses/MIT>.
+///
 
 #ifndef soundbox_sounds_hpp
 #define soundbox_sounds_hpp
 
+#include <string>
+
 std::string play_sound();
-void getAudioDevice();
+std::string getAudioDevice();
 
 #endif
