+++ win_samples/RenderExclusiveEventDriven/ToneGen.h
