+++ win_samples/RenderExclusiveEventDriven/WASAPIRenderExclusiveEventDriven.cpp
