+++ win_samples/RenderExclusiveEventDriven/WASAPIRenderExclusiveEventDriven.sln
