+++ win_samples/RenderExclusiveEventDriven/WASAPIRenderExclusiveEventDriven.vcproj
