+++ win_samples/RenderExclusiveEventDriven/WASAPIRenderer.cpp
