+++ win_samples/RenderExclusiveEventDriven/WASAPIRenderer.h
