+++ win_samples/RenderExclusiveEventDriven/stdafx.cpp
