+++ win_samples/RenderExclusiveEventDriven/stdafx.h
