+++ win_samples/RenderExclusiveEventDriven/targetver.h
